refactor: rename wait to cadastro-finalizado
